Refactor: separate psbt struct from lib.rs
